Update StockController removed dynamic implementation.
Dynamic implementation goes against the principle of having types in the first place.

Solution currently isolates the code to fetch a list of all StockManage entries, while keeping the previous type declarations.

Integrating search functionality is going to be slightly more complicated now.

Given a scenario where we want to query stock from supplier 1 for the past 30 days, the parameters have to be available to GetAllStock. If we were to generate a csv file after every query, and keep the code within Index, then it is likely that the system will do extra work for no reason.